services/adapters/k8s_adapter/requirements.txt
pyyaml==6.0.1
asyncio-subprocess==0.1.0